moved Results class to pipeline module
